Fixing dtype and device issues

+++ src/nak_torch/algorithms/cbs.py
@@ -29,7 +29,14 @@ def cbs_step(
     drift_term = particles_diff.neg_()
     noise_sqrt_cov = sym_sqrtm(particles_cov.mul_(motion_scaling_sq))
     motion_term = (
-        torch.normal(0.0, 1.0, particles.shape, generator=rng, device=rng.device)
+        torch.normal(
+            0.0,
+            1.0,
+            particles.shape,
+            generator=rng,
+            dtype=particles.dtype,
+            device=rng.device,
+        )
         @ noise_sqrt_cov
     )
     return drift_term, motion_term
